+++ reflect.go
@@ -67,24 +67,22 @@ func derefPointers(rv reflect.Value) (drv reflect.Value, isPtr bool, isNilPtr bo
 	return
 }
 
-// Returns isNil=true iff is ultimately nil after (recursive) dereferencing.
-// If isNil=false, erv is set to the non-nil (valid) dereferenced value.
-// Empty non-pointers/non-interfaces or 0-length slices are not nil.
-// 0-length slices will return isEmpty=true.
-func isNilOrEmptySafe(rv reflect.Value) (erv reflect.Value, isNil bool, isEmpty bool) {
+// Returns isVoid=true iff is ultimately nil or empty after (recursive) dereferencing.
+// If isVoid=false, erv is set to the non-nil non-empty valid dereferenced value.
+func isVoid(rv reflect.Value) (erv reflect.Value, isVoid bool) {
 	rv, _, isNilPtr := derefPointers(rv)
 	if isNilPtr {
-		return rv, true, false
+		return rv, true
 	} else {
 		switch rv.Kind() {
-		case reflect.Slice:
-			isNil := rv.IsNil()
-			isEmpty := !isNil && rv.Len() == 0
-			return rv, isNil, isEmpty
-		case reflect.Chan, reflect.Func, reflect.Interface, reflect.Map:
-			return rv, rv.IsNil(), false
+		case reflect.String:
+			return rv, rv.Len() == 0
+		case reflect.Chan, reflect.Map, reflect.Slice:
+			return rv, rv.IsNil() || rv.Len() == 0
+		case reflect.Func, reflect.Interface:
+			return rv, rv.IsNil()
 		default:
-			return rv, false, false
+			return rv, false
 		}
 	}
 }